notebooks/wxO_api/requirements.txt
Pillow
botocore
docling
docling_core
elasticsearch
ibm-cos-sdk
langchain
langchain_community
langchain_core
pandas
requests
urllib3 